Jud Participates
Here I have done a bit of code cleaning and documentation.
I like that you've messed with this a bit. I think if you are
generally on board we need to hammer down why we are doing this
and what we can do to accomplish our clearly defined objectives.
I think this is a great idea to tinker with. I have provided my
desired directives in the documentation. Feel free to expand/contract
what I have provided.

Anywho...
