+++ .gitignore
@@ -1,2 +1,3 @@
 # Put your Confluence username, API key, and domain in this.
 .env
+pubdoc
